backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Werkzeug==3.1.3
Pillow==11.1.0
APScheduler==3.11.0
python-dotenv==1.0.1
bcrypt==4.2.1
